keystone: tci6638: add I2C base address defines
This patch adds the defines of the base address of the I2C0, I2C1
and I2C2.  The 512Kb EEPROMs are on I2C0 while the DDR3A EEPROM is
on I2C1.  Currently u-boot supports only one I2C. On Kepler EVM,
I2C0 is the default supported I2C.  If I2C1 needs to be supported
for reading the DDR3A SPD, redefine I2C_BASE to I2C1_BASE.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>