add nand ecclayout command.
This commit adds a nand ecclayout command that allows the ecclayout of
the current nand device to be changed during run time.  This feature is
useful when using u-boot to write something to nand flash that will be
read by other applications, such as ROM bootloader, that expects a
differnet ECC layout.  In that case, change the current nand device
ecclayout using the "nand ecclayout set" command before writing the
data to nand flash.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>