usb/gadget: add the fastboot gadget
This patch contains an implementation of the fastboot protocol on the
device side and a little of documentation.
The gadget expects the new-style gadget framework.
The gadget implements the getvar, reboot, download and reboot commands.
What is missing is the flash handling i.e. writting the image to media.

Signed-off-by: Sebastian Andrzej Siewior <bigeasy at linutronix.de>

From u-boot mailing list:
http://lists.denx.de/pipermail/u-boot/2011-September/101557.html