keystone: net: add helpers for POST
POST EMAC PHY loop-back test requires having the mdio to set
the phy in loop-back. But the driver doesn't need to checl the phy
during that test.

This patch add the loopback_test variable. If the variable is set,
driver doesn't access the phy.

The patch adds helper functions to set the loopback_test variable and
get the value of the sys_has_mdio variables.

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>