tci6638: passing more than 2GB ddr3a memory size to kernel
U-boot works without MMU enabled and therefore cannot detect and use
more than 2GB of DDR3A. But u-boot has to pass the correct memory
configuration to kernel. It does that by fixing the memory node in
the device tree.

Even if the DDR3A physically is one solid bank, u-boot uses two logical
banks. The first bank is configured in the same way as before.
Thus it maximum size is 2GB and it may be reduced by the size of
reserved memory. We reserve memory at the end of the first 2GB of
DDR3A as that is only memory accessible from DSPs.

If the "mem_lpae" environment variable is set and the "ddr3a_size" is set
to 4 or 8 (GB) u-boot creates the second memory bank with the size equal
to (ddr3a_size - 2).

Signed-off-by: Vitaly Andrianov <vitalya@ti.com>