tci6638: workaround for ddr3a/3b memory issue
This patch implements a workaround to fix DDR3 memory issue.
the code for workaround detects PGSR0 errors and then preps for
and executes a software-controlled hard reset.	In board_early_init,
where logic has been added to identify whether or not the previous
reset was a PORz. PLL initialization is skipped in the case of a
software-controlled hard reset. The workaround is in sync with
v1.5 of the Gel file

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Keegan Garcia <kgarcia@ti.com>